feat: replace ROM by CRT, make video mode switch in freezer #31
as ROM load wasn't working anyway and MEGAINFO is the better way to see
all versions, I replaced ROM with C(R)T Emu. This will switch CRT
emulation on and off (live)
I also changed the video mode switch so that it will switch the freezer
itself in addition to the frozen system.